fetching owner id from jwt to delete rooms
created delete api for rooms

+++ src/main/java/com/stayvida/backend/dto/LoginResponse.java
@@ -3,17 +3,16 @@ package com.stayvida.backend.dto;
 public class LoginResponse {
     private boolean success;
     private String token;
-    private Long id;
+    private int id;
     private String email;
     private String role;
     private String message;
     private boolean profileExists;
 
-
     public LoginResponse() {
     }
 
-    public LoginResponse(boolean success, String token, Long id , String email,  String role, String message) {
+    public LoginResponse(boolean success, String token, int id, String email, String role, String message) {
         this.success = success;
         this.token = token;
         this.id = id;
@@ -30,28 +29,29 @@ public class LoginResponse {
     public String getToken() {
         return token;
     }
-    public Long getuserID() {
+
+    public int getuserID() {
         return id;
     }
-
 
     public String getEmail() {
         return email;
     }
 
-       public String getRole() {
+    public String getRole() {
         return role;
     }
-
 
     public String getMessage() {
         return message;
     }
+
     public boolean isProfileExists() {
         return profileExists;
-    }   
+    }
+
     public void setProfileExists(boolean profileExists) {
         this.profileExists = profileExists;
-    }    
- 
+    }
+
 }
